moving dcfglib into pkg as dcfg, also fixing tests
